ReservedFunctionNames: refactor the sniff and use PHPCSUtils
Refactor the sniff to be stand-alone, independent of the PHPCS sniff.

The sniff only extended the PHPCS sniff for the properties containing the names of magic methods and such anyway and these can now be found, including helper methods, in the `PHPCSUtils\Utils\FunctionDeclarations` class, where the function name lists are also more complete.