ForbiddenStripTagsSelfClosingXHTML: use PHPCSUtils
1. Use `BCTokens::textStringTokens()`

This is a property which was introduced in PHPCS 2.9.0, so is not available to us natively, but PHPCSUtils backfills it.

2. Cleaner error output

The PHPCSUtils `PassedParameters::getParameters()` method provides both a `raw`, as well as a `clean` index, where `clean` contains the token content stripped of comments.

Using the `clean` index will allow for "cleaner" code snippets in the error messages.