NewListReferenceAssignment: use PHPCSUtils / refactor the sniff
As PHPCS has an issue with sniffs extending other sniffs, it is better for sniffs to be stand-alone.

PHPCSUtils contains the `Lists::getAssignments()` method. Using that method takes all the heavy lifting away from this sniff, so it is now easy to refactor the sniff to be stand-alone without having lots of duplicate code.