NewUniformVariableSyntax: bug fix - false negative when used with hierarchy keywords
I ran some tests as I thought it was strange that `T_PARENT` wasn't accounted for and I intended to replace the check for the `T_SELF` and `T_STATIC` token with one using `Collections::$OOHierarchyKeywords`.

My findings were contrary to what was documented in-line: function calls with the hierarchy keywords, double colon and then a variable are also affected by the change in PHP 7.0.

I tried to find in the history of the sniff why I made an exception for this previously and also looked over the RFC, but I haven't been able to determine why the exception was previously made, so I'm removing it now.

Proof of false negative: https://3v4l.org/umg8a

Includes adjusted unit tests.