RemovedPredefinedGlobalVariables: bug fix for variables imported via closure `use`
This was a known bug but would previously have required a lot of custom code to work out.

Now, using the `FunctionDeclarations::getParameters()` which supports getting the parameters from a closure `use` import, checking this has become a lot easier.

If a closure imports the `$php_errormsg` variable in a `use` statement, the use of this variable within the closure will be regarded as ok. The use of the variable in the `use` itself will still check the surrounding scope for the variable being declared.

Includes adjusting the unit tests for the fixed bug.