ForbiddenNames: use `inc` extension for the test case files
The test case files for the `ForbiddenNames` sniff were the only ones using a `php` extension instead of a `inc` extension.
This meant that exceptions for these files had to be made in both the Composer `lint` command, as well as in the `phpcs.xml.dist` file.

This commit fixes that by changing the file extension used for the test case files for the `ForbiddenNames` sniff to `inc`.