ForbiddenNames: add missing keywords
There were still a few reserved keyword missing from the list, even after 923, which fixed most of this.

As this means that the base test now will include a keyword (`fn`) which is PHP 7+, the method name/class constant tests have to be split off to still allow for testing that these _do_ trigger an error when PHP 5 needs to be supported, but _don't_ trigger an error when the minimum version is PHP 7.

Includes regenerated unit test case files.
Includes adding test code for the "valid keyword uses" test case file.