ForbiddenNames: improve import use statement aliases check
This abstracts out the check on import use statements to its own method and makes sure that multi-use statements and group use statements are handled correctly.

This includes removing the check on names being imported without alias.
These cannot be changed until the declaration is changed.
Also: as names are normally multi-part, this check was incorrect anyways as it only checked the first token of the name.

Note: two errors are thrown for the `continue` and the `global` keywords being used as an alias.
This is due to a "bug" in PHPCS where the semi-colon directly following a `const` or `function` keyword in a `use` statement is tokenized as `T_STRING`, not `T_SEMICOLON`.
As this type of code is a parse error anyway, this is not something which really ought to be fixed in PHPCS and working around it in PHPCompatibility seems superfluous. After all, the error _is_ being thrown and the duplication is due to a parse error in the provided code.

Example code snippet:
```php
use Foobar as Const; // <- This semicolon is incorrectly tokenized.
use Foobar as Continue;
```