ForbiddenNames: merge the `ForbiddenNamesAsDeclared` sniff [1]
The logic for the checks in the `ForbiddenNamesAsDeclared` sniff is largely the same as the logic for `ForbiddenNames`, so merging these sniffs reduces code duplication.

Also: the `ForbiddenNamesAsDeclared` sniff did not check for the "other" reserved keywords being used as aliases in import `use` statements. By merging the sniffs, this will now be checked for as well.

This commit adjusts the tests to allow for the merge:
* Adds test cases for valid uses of the "other" reserved keywords to the `ForbiddenNamesUnitTest.1.inc` test case file.
* Adjusts the script for auto-generating the main test cases files to selectively include the "other" reserved keywords.
* Adjusts the unit test file to selectively expect the errors for the "other" reserved keywords and to test against false positives.

Includes regenerated test case files.