PHP 8.0: RemovedPHP4StyleConstructors - handle removed support in PHP 8
> Methods with the same name as the class are no longer interpreted as
constructors. The __construct() method should be used instead.

Refs:
* https://github.com/php/php-src/blob/69888c3ff1f2301ead8e37b23ff8481d475e29d2/UPGRADING#L24
* https://github.com/php/php-src/commit/682b54f68748715f85e9ac4a267477d9ac61918a

This commit changes the existing sniff to throw an `error` instead of a `warning` if the `testVersion` includes PHP 8 and adjusts the error message to reflect this.

Note: the error code has also been adjusted. Previously, it would be `Found`, now it will be `Deprecated` or `Removed`, depending on the `testVersion` set and whether a warning or error is being thrown.
:point_right: This needs a changelog entry.

Includes adjusted unit tests.