PHP 8.0: RequiredToOptionalFunctionParameters - handle optional param for three GD functions
> The $num_points parameter of imagepolygon(), imageopenpolygon() and
>  imagefilledpolygon() is now optional, i.e. these functions may be called
>  with either 3 or 4 arguments. If the arguments is omitted, it is calculated
>  as count($points)/2.

Implementation note:
The `$num_points` parameter is the 3rd parameter and has become optional, but the `$color` parameter which is the 4th parameter, is still expected to be passed, so detection needs to take place on the minimum number of arguments having changed, not on the actual position of the argument now optional.

Refs:
* https://github.com/php/php-src/blob/69888c3ff1f2301ead8e37b23ff8481d475e29d2/UPGRADING#L611-L614
* https://github.com/php/php-src/commit/2de79f085502315d1f53822eee54e9f39dd809f6

Includes adjusted unit tests.