Tests: make the test suite compatible with PHPCS 4.x
The underlying PHPCS native method to set config data is no longer static as of PHPCS 4.x, which means the existing `$config` as required as input for the `Helper` method in PHPCS 4.x.

Refs:
* PHPCSStandards/PHPCSUtils 137
* https://github.com/squizlabs/PHP_CodeSniffer/commit/10a89a2c8c33a26466ccb3368a69a5e99549f80e

Note: there are still other changes needed to the sniffs to make them compatible with PHPCS 4.x. This PR only addresses the changes needed to the test suite.