NewTypedProperties: refactor to use the AbstractNewFeature sniff as a base
As PHP 8 introduces new "types", the `NewTypedProperties` sniff needs the ability to throw the generic "Typed properties are not supported" error, as well as the ability to throw specific "This type was not supported before PHP x" errors for newly introduced _types_.

Up to now, only the first type of error message was supported.

This commit refactors the sniff to extend the `AbstractNewFeature` sniff to allow that sniff to handle the complex version number array negotiations.

With this initial refactor, the new `$newTypes` property is still empty, but that will change in the near future when more PHP 8 features are pulled to the repo.