NewTypedProperties: move invalid types to property
The sniff as is, contains a cursory check for invalid types.

This commit moves the list of invalid types into a property to allow for more easily extending the list.

It also adds `boolean` and `integer` to the list, just like in the `NewParamTypeDeclarations` sniff, and allows for an alternative to be suggested.

Note: `static` has not been added to the list of invalid types as it is a property keyword and it wouldn't be possible to distinguish whether it was used as a keyword or a type (which is also why it isn't supported in PHP8).

Includes unit tests for the newly added types.