PHP 8.0: NewTypedProperties: allow for "mixed" type
As of PHP 8.0, `mixed` can be used as a type for property declarations.

The RFC explicitly does not allow for the `mixed` type to be combined with the nullability indicator.
While in itself, that is not a cross-version compatibility issue, I have made a very conscious choice to add a check for this anyway as - while discouraged with `mixed` being a soft reserved keyword -, prior to PHP 8, a class _could_ be named `mixed`, so a `?Mixed` type hint referring to such a class could exist in code and would be a cross-version compatibility issue as since PHP 8 that will throw a "Fatal error: Mixed types cannot be nullable".

Refs:
* https://wiki.php.net/rfc/mixed_type_v2
* https://github.com/php/php-src/pull/5313
* https://github.com/php/php-src/commit/aec4c0fd031ad557527ff6c888a1b60048bb1cdc

Includes unit tests.