NewReturnTypeDeclarations: change the error position for class name based types
In contrast to the other errors, class names can have multiple tokens and the error for a class name based return type would be thrown on the _last_ (class name) token of the type declaration, not the first (and only) as was done for the other types.

This minor change, syncs the error position for all return type errors to the _start_ of the return type.