NewTypedProperties: saner unit test organization
The "no violation" test was currently organized as part of the "invalid property type" tests.

This is not strictly correct. It's just that the "no violation" test can not be executed for the "invalid property type" test cases.

What with more property types being added in PHP 8 as well as other features, this will become true of more test cases, not related to them being an "invalid property type".

So for sanity sake, I'm moving the "no violation" test to the "is this is typed property" test, with a toggle to turn that test off for test cases which would also throw errors in later PHP versions.