PHP 8.0: account for removed extension XMLRPC
The [XMLRPC](https://www.php.net/manual/en/book.xmlrpc.php) extension will be removed without deprecation in PHP 8.0.

This commit adds the functions and class from the extension to the respective `RemovedFunctions` and `RemovedClasses` sniffs.

The class is not listed in the PHP manual, but is in the stub file (as a final class), so, while I'm not 100% sure, I do think we should account for it.

Includes unit tests.

Refs:
* https://wiki.php.net/rfc/unbundle_xmlprc
* https://github.com/php/php-src/commit/4fa3687e56050a111e8eaffe330258003628225e
* https://www.php.net/manual/en/book.xmlrpc.php
* https://github.com/php/php-src/blob/444aa2636e81053e439623f2dbf981eaa48e6161/ext/xmlrpc/xmlrpc.stub.php