RemovedGetDefinedFunctionsExcludeDisabledFalse: minor tweak for consistency
Oops... just saw this after PR 1150 was merged ...

The feature is - at this time - only deprecated. Everywhere else, we throw a "warning" for deprecations and an "error" for removals/new features, so this should have been set to throw a "warning" for now.

Also, as at some point in the future, the feature is expected to be removed, let's change the error code to `Deprecated` to allow for a second `Removed` error code once the feature is removed.
Again, that is in line with how it's done in most places in the code base (though a cross-sniff consistency check at some point wouldn't be amiss).
Having different error codes for Deprecated/Removed allows for ignoring deprecations, while still getting the errors when the feature actually gets removed.