RemovedNamespacedAssert: bug fix - handling of nested function declarations / use PHPCSUtils
A function called `assert()` declared nested in a class method or closure in a namespaced file will still be defined in that namespace and is therefore deprecated.

POC:
* https://3v4l.org/fKBkf
* https://3v4l.org/mMXLv

This fixes the handling of those.

Includes out the PHPCS native `File::getDeclarationName()` method for the PHPCSUtils `FunctionDeclarations::getName()` method.

Includes unit tests.