Various sniffs: use Collections::functionDeclarationTokensBC()
PHPCSUtils 1.0.0-alpha3 introduced a number of new token group properties and methods, including the `Collections::functionDeclarationTokensBC()` method which returns all tokens which can be used to declare functions, i.e. `T_FUNCTION`, `T_CLOSURE`, `T_FN` (if it exists) and `T_STRING` for when arrow functions have not been backfilled yet.

This implements the use of this method in various sniffs which have already previously been reviewed & updated for PHPCSUtils.