NewStripTagsAllowableTagsArray: use PHPCSUtils
1. Use `BCTokens::textStringTokens()`

This is a property which was introduced in PHPCS 2.9.0, so is not available to us natively, but PHPCSUtils backfills it.

2. Improve comment tolerance

The PHPCSUtils `PassedParameters::getParameters()` method provides both a `raw`, as well as a `clean` index, where `clean` contains the token content stripped of comments.

This allows for more reliable and faster results for the sniff.

Includes unit test.