RemovedIconvEncoding: bug fix / compatibility with PHP 8
The `$data` parameter of the PHPCS `addError()`/`addWarning()` methods should always be passed as an array.

On PHP < 8, if it wasn't PHP was tolerant and accepted it anyway (cast to an array). On PHP 8.0, this is no longer the case and not passing `$data` as an array will result in a fatal `TypeError`.