PHP 8.0 | NewTrailingComma: add check for trailing comma in closure use lists
PHP 8.0 introduced trailing comma's in function declaration parameter lists. A sniff to detect these was added via PR 1164.

Since then, a new RFC has been voted on & accepted, to also allow trailing comma's in closure use lists.

This adds an additional check to the `PHPCompatibility.FunctionDeclarations.NewTrailingComma` sniff to detect trailing comma's in closure `use` lists.

Note: the error code for the original error message has been changed from `Found` to `InParameterList`, but as this sniff was only introduced recently and has not been released yet, this is not an BC-break.

Refs:
* https://wiki.php.net/rfc/trailing_comma_in_closure_use_list
* https://github.com/php/php-src/pull/5793