Travis: fix build failure
For some unfathomable reason, the quicktest build on PHP 7.3 against PHPCS 2.6.0 passes, while the coverage build for the same fails on the null coalesce equals operator test in the `NewOperators` sniff.

The only real difference between the two builds is the addition of the Coveralls dependency, but even looking at the dependencies, it its still unclear why this makes a difference.

As PHPCS 2.6.0 isn't fully compatible with PHP 7.3 anyhow (continue in switch warnings) and there is a build for that PHP/PHPCS combination in the quick test stage, I'm downgrading the PHP version for the coverage build to PHP 7.2, which does pass.