CI: fix coverage builds failing
The code coverage builds on PHP 7.3/7.4 started to fail without clear reason recently.

It has now become clear that the underlying reason is that the [PHP Code Coverage](https://github.com/sebastianbergmann/php-code-coverage) package started to use PHP-Parser as of PHP Code Coverage version 9.0.

The first PHPUnit version which has a dependency on PHP Code Coverage 9.x is PHPUnit 9.3, so if we limit the PHPUnit 9.x range to `9.0 - 9.2`, the code coverage builds will be fine (for now).

We then have the situation that PHPUnit 9.3 is needed to run the tests on PHP 8.0. As we don't run code coverage on PHP 8.0, that is not a really issue though, so we just hard-require PHPUnit 9.3 for PHP `nightly` in the Travis script.

This now also allows to revert the work-around for the PHP 7.3 code coverage build failing as was put in place in 63abdd8cd037322fa58e908f64139e1db5c07659.

While this solution is not ideal, it at least will create a workable situation again until PHP 8.0 comes out. Hopefully a proper solution will be found and implemented before that time, so we can remove this work-around.