:wrench: Travis/Composer: use PHPCSDevCS for the code-style check
PHPCSDevCS is an external standard for PHPCS which is largely based on the PHPCS project ruleset which was already in use by PHPCompatibility with a few extra tweaks, including switching from PSR2 to PSR12.

The `PHPCSDevCS` package has a minimum PHPCS requirement of PHPCS 3.5.0 to take advantage of the latest sniffs.

Now, there are a few challenges with using the package:
1. The above mentioned PHPCS 3.5.0 requirement which would make testing with lower PHPCS versions difficult if we'd `require-dev`-ed the package.
2. A circular dependency as PHPCompatibility is a dependency of `PHPCSDevCS` - and we want the PHPCompatibility version to be used to be the current branch.
3. The fact that we are currently working on a new major, which means that, while the latest stable release is `9.x`, we already want to be using `10.x`.

To get round that, a couple of things are needed:
0. `PHPCSDevCS` needs to allow both PHPCompatibility `9.x` as well as `10.x` to be installed.
    This was added in version `1.1.1` of `PHPCSDevCS`.
    - `9.x` for other packages, like PHPCSUtils, using the standard until we release `10.0.0`.
    - `10.x` for us.
1. Branch aliases for the `master` and `develop` branches in the `composer.json` file.
    With these in place, running `composer install` locally and then requiring `phpcsstandards/phpcsdevcs` will work fine.
2. Setting `COMPOSER_ROOT_VERSION` in the Travis CI script as Travis doesn't do a full clone of the repo, so would still not be able to determine the version to use for the active branch correctly.

Note:
* The `branch-alias` for `dev-master` will need to be updated before we release version `10.0.0`.
* The `branch-alias` for `dev-develop` and the version`set as `COMPOSER_ROOT_VERSION` will need to be updated when work would start on a new major version.

Last thing - as the PHPCS requirements for `PHPCSDevCS` and `PHPCompatibility` are different, we don't want the package to be in `require-dev` as it would make testing more difficult for people using a Composer based dev setup.

So, instead, I've added Composer scripts to add the standard only for CS runs and remove it afterwards.

Refs:
* https://github.com/PHPCSStandards/PHPCSDevCS
* https://twitter.com/ramsey/status/1299455790823215105

With graceful thanks to ramsey, Seldaek and naderman for helping me figure this all out.
(and not just for this repo, similar PRs will be lined up for some other repos using the standard in the near future)