:wrench: CS/QA: switch over to using the `PHPCSDev` ruleset
The `PHPCSDev` ruleset is largely based on the custom `phpcs.xml.dist` ruleset which was already in use for the PHPCompatibility standard.

By having `PHPCSDev` available as a named ruleset, we can more easily enforce consistency in the codestyle used for the new repos within the PHPCompatibility organisation (as well as the ones in PHPCSStandards, like PHPCSUtils).

This switches out the custom ruleset in favour of `PHPCSDev` with only a few exceptions.

Ref: https://github.com/PHPCSStandards/PHPCSDevCS