PHP 8.0 | OCI8: account for renamed classes
> The OCI-Lob class is now called OCILob, and the OCI-Collection class is now
> called OCICollection for name compliance enforced by PHP 8 arginfo
> type annotation tooling.

This commit:
* Adds the new classes to the `NewClasses` sniff.
* Updated the `alternative`s for various deprecated/removed functions in the `RemovedFunctions` sniff.

Based on the commit, it looks like the _old_ classnames are removed without deprecation. Making a note of this in the `RemovedClasses` sniff, but not enabling the check as the places where we are searching for the name, do not allow for the old name in the first place - would have been a parse error. Also see: https://bugs.php.net/bug.php?id=79625

Refs:
* https://github.com/php/php-src/blob/0a84fba0deb1c1b75770a436c4236dc56e6d0463/UPGRADING#L392-L394
* https://github.com/php/php-src/commit/ff66e4945ef1936a02ea149e03925a9e01b606e0