PHP 8.0 | OptionalToRequiredFunctionParameters: handle crypt $salt becoming required
> Calling crypt() without an explicit salt is no longer supported. If you
>  would like to produce a strong hash with an auto-generated salt, use
>  password_hash() instead.

The `$salt` parameter for `crypt()` was previously _strongly recommended_ to be passed, but not passing it was not formally deprecated.

With this in mind, the sniff special-cased this check.

With the parameter now being required, building the error message in combination with the special casing of "recommended" instead of "deprecated" would become exponentially more complicated.

With this in mind, I've removed the special-casing and, while not _technically_ correct, with a `testVersion` including PHP 5.6, the optional nature of the parameter will be reported as _deprecated_ instead of _recommended_.

This includes a change in the errorcodes being used by the sniff, which should be annotated in the changelog.

Refs:
* https://github.com/php/php-src/blob/0a84fba0deb1c1b75770a436c4236dc56e6d0463/UPGRADING#L608-L610
* https://github.com/php/php-src/commit/032f862133dbd2acc04cb75004428d6209f6046b