ArgumentFunctionsReportCurrentValue: don't report when parameter is only used in simple assignment
Prevent a false positive (warning) when a parameter is only _assigned_ to another variable and this is a "simple assignment".

For the purposes of this check, a "simple assignment" is defined as:
- The variable is not nested in parenthesis, i.e. not used in a potential function call which may adjust the variable by reference.
- Not preceded by a reference operator (reference assignment).
- Has an assignment operator before it and none after.

Any more complex statements where a passed parameter is being assigned to another with a call to one of the argument functions after it, will still be reported as "needs inspection".

Includes additional unit tests.

Fixes 1240

Related to 796