Tests: Xdebug 3 compatibility fix
We are currently limiting the PHPUnit version being used to max PHPUnit 9.2.x due to an issue with the code coverage package having switched over to `PHP_Parser` which defines token constants for constants which the sniffs rely on not to exist unless a PHP_CodeSniffer version is used in which they are backfilled.

However, Xdebug `3.0` was released last week and the Travis images for high PHP versions have been updated to include Xdebug `3.0`, but now it turns out that PHPUnit 9.2.x is not compatible with Xdebug 3....

Declaring the "missing" constants which no longer exist in Xdebug 3, but which PHPUnit 9.2 is relying on to exist, will fix this, combined with enabling the Xdebug code coverage mode from within the Travis config.