NewConstantArraysUsingDefine: add support for named parameters
PHPCSUtils `1.0.0-alpha4` will add support for named parameters.
Ref: https://github.com/PHPCSStandards/PHPCSUtils/pull/235

To allow for this, any calls to `PassedParameters::getParameter()` to retrieve a parameter in a function call need to pass the name of the target parameter.

Within PHPCompatibility, this is the only sniff for which the tests would fail at the moment when used with PHPCSUtils alpha-4, so pre-eminently fixing this ahead of the release of PHPCSUtils 1.0.0-alpha4.

Adding unit tests for this change would break on alpha3 though, so I'm leaving those for a future PR once alpha4 has been released.