Ini sniffs: move `$iniFunctions` property to sniffs
As this property is only used by two sniffs, having it declared in the individual sniffs instead of in the base `Sniff` class will simplify things when the sniffs will change to a different parent sniff (once the new parent becomes available in PHPCSUtils).

Doing this now, prevents us having to deprecate the property in a minor and removing it in the next major.