CI: switch to GitHub Actions - fix uploading coverage run results
Uploading the test coverage `clover.xml` file to Coveralls works with PHP Coveralls 2.4+, but not with PHP Coveralls 1.x, but version 2.x is only compatible with PHP 5.5+, while we are running some coverage builds on PHP 5.4.

Luckily a "hot swap" of the PHP version during a job turns out to work just fine, so adding steps to do so, which allows us to only use PHP Coveralls 2.x for the executing the upload to Coveralls.