Contributing: add links to PHP 4 and PHP 5 documentation
The official PHP manual is removing references to PHP 4 and PHP 5.

For the purposes of PHPCompatibility, it is important to still be able to have access to that documentation to find details of changes and old vs new behaviour.

This commit adds links to a mirror which will keep copies of the PHP 4 and PHP 5 documentation available for reference.

Note: links to the PHP manual used in the docblocks of sniff classes may no longer work now the PHP 4 and 5 documentation is being removed and some may need to be changed as well. A separate issue will be opened for that.