Refactor Complex version sniffs: introduce traits
In anticipation of abstract sniffs to detect the use of named classes, interfaces, functions in PHPCSUtils in the near future, the sniffs using complex version arrays are being refactored.

This is done now, even though the abstract sniffs are not yet available in PHPCSUtils, as doing so later would be a BC-break and would necessitate a new major release, while doing so now allows for us to implement the abstract sniffs from PHPCSUtils in a future minor release.

The original abstract sniffs for complex version arrays as contained in PHPCompatibility were not feature detection based, but based on the type of complex version array: new vs deprecated/removed.

As the abstract sniffs in PHPCSUtils will be feature detection based, this commit abstracts the logic related to the complex version arrays out to traits which can be used in combination with the abstract sniffs in PHPCSUtils.

Some of the logic currently contained in the PHPCompatiblity abstract sniffs will be moved to the actual sniffs.

Includes dedicated unit tests for the traits.