PHP 8.1 compatibility: fix deprecation notice
PHP 8.1 deprecates passing `null` to PHP native functions where the parameter(s) is not explicitly nullable.

While in PHP 8.1 this is only a deprecation, it should, of course, still be fixed.

This commit contains a fix for the one issue in PHPCompatibility found related to this so far (based on the unit tests).

Passing `null` to `trim()` is now deprecated, so more defensive coding or explicit type casting is needed.
This comes into play for the `TestVersionTrait::getTestVersion()` method.

Includes removing a stray line of commented out code.