RemovedCallingDestructAfterConstructorExit: improve warning message
This sniff can throw an error or a warning depending on whether a `destruct()` method has been found in the class (error) or not, but the class has a parent class or uses traits (warning).

This improves the message shown for the `warning` case to make it clearer how to take action on the issue.

It also ensure that the error and the warning each have a unique error code.