ArgumentFunctionsReportCurrentValue: don't report when parameter is only used in return/exit [2]
Follow up on 1208.

Improve the previously applied fix to bow out from reporting "use" of a passed parameter before a call to `func_get_args()` when the parameter is used in a `return` or `exit` statement.

The previous fix would break when the `return`/`exit` statement was within an unscoped condition.

Unscoped conditions (no curly braces) are a bane to the existence of sniff writers anywhere and can't be fully accounted for.
Even so, there are some common situations in which the previous fix can be improved to reduce false positives.

This current fix will skip over `return`/`exit` statements completely, instead of walking back from a matched variable to check if something is within a `return`/`exit` statement.
This also makes the sniff more efficient.

Secondly, the sniff will now also ignore parameter use in `throw` statements.

Includes additional unit tests.

Fixes 1240