Composer: tweak the PHPUnit 4.x version used
The PHPUnit 4.x version used was based on a PHPUnit bug previously encountered in 511.

It appears though that with changes which have been made to the test set up in the mean time, this bug no longer rears its ugly head, so we can now switch back to using the highest available PHPUnit 4.x version again.